requirements.txt
appdirs==1.4.4
black==20.8b1
click==7.1.2
mypy-extensions==0.4.3
pathspec==0.8.0
psycopg2-binary==2.8.6
python-dotenv==0.14.0
regex==2020.7.14
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
